Add Auth for journal end points

+++ src/main/java/journalApp/service/JournalEntryService.java
@@ -33,7 +33,7 @@ public class JournalEntryService {
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
 //            user.setUserName(null);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e){
             log.error("Exception",e);
             throw new RuntimeException("An error occurred while saving the entry.",e);
@@ -49,11 +49,21 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed =false;
+            try {
+                User user = userService.findByUserName(userName);
+                 removed =user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+                if(removed){
+                    userService.saveUser(user);
+                    journalEntryRepository.deleteById(id);
+                }
+            }catch (Exception e) {
+                System.out.println(e);
+                throw new RuntimeException("An Error occured while delete the entry",e);
+            }
+            return removed;
     }
 
 }
